+++ src/jsoninfo.hh
@@ -9,8 +9,8 @@
 
 struct jsoninfo {
     nlohmann::json* j;
-    int doc_start;
-    int header_start;
+    long int doc_start;
+    long int header_start;
     unsigned int allocated_size;
     std::string name;
 };
